+++ src/docker/init.js
@@ -10,6 +10,7 @@ const docker = require('./docker');
 const logger = require('../logger');
 const {initNetwork} = require('./network');
 const {getPlugins} = require('../plugins');
+const {initTraefik} = require('./traefik');
 
 // config vars
 const baseFolder = path.join(os.homedir(), '.exoframe');
@@ -70,20 +71,6 @@ exports.initDocker = async () => {
   // get config
   const config = getConfig();
 
-  // check if traefik management is disabled
-  if (!config.traefikImage) {
-    logger.info('Traefik managment disabled, skipping init..');
-    return;
-  }
-
-  // build acme path
-  const acmePath = path.join(baseFolder, 'traefik', 'acme');
-  try {
-    fs.statSync(acmePath);
-  } catch (e) {
-    mkdirp.sync(acmePath);
-  }
-
   // run init via plugins if available
   const plugins = getPlugins();
   logger.debug('Got plugins, running init:', plugins);
@@ -101,102 +88,6 @@ exports.initDocker = async () => {
     }
   }
 
-  // get all containers
-  const allContainers = await docker.listContainers({all: true});
-  // try to find traefik instance
-  const traefik = allContainers.find(c => c.Names.find(n => n.startsWith(`/${config.traefikName}`)));
-
-  // if traefik exists and running - just return
-  if (traefik && !traefik.Status.includes('Exited')) {
-    logger.info('Traefik already running, docker init done!');
-    return;
-  }
-
-  // if container is exited - remove and recreate
-  if (traefik && traefik.Status.startsWith('Exited')) {
-    logger.info('Exited traefik instance found, re-creating...');
-    const traefikContainer = docker.getContainer(traefik.Id);
-    // remove
-    await traefikContainer.remove();
-  }
-
-  // pull image if needed
-  const allImages = await docker.listImages();
-  const traefikImage = allImages.find(img => img.RepoTags && img.RepoTags.includes(config.traefikImage));
-  if (!traefikImage) {
-    logger.info('No traefik image found, pulling..');
-    const pullLog = await pullImage(config.traefikImage);
-    logger.debug(pullLog);
-  }
-
-  // debug flags
-  const debug = ['--debug', '--logLevel=DEBUG'];
-
-  // letsencrypt flags
-  const letsencrypt = [
-    '--acme',
-    `--acme.email=${config.letsencryptEmail}`,
-    '--acme.storage=/var/acme/acme.json',
-    '--acme.httpchallenge.entrypoint=http',
-    '--acme.entrypoint=https',
-    '--acme.onhostrule=true',
-    '--accesslogsfile=/var/acme/access.log',
-    `--entryPoints=Name:https Address::443 TLS ${config.compress ? 'Compress:on' : 'Compress:off'}`,
-    `--entryPoints=Name:http Address::80 Redirect.EntryPoint:https ${config.compress ? 'Compress:on' : 'Compress:off'}`,
-    '--defaultEntryPoints=https,http',
-  ];
-
-  // entrypoints without letsencrypt
-  const entrypoints = [
-    `--entryPoints=Name:http Address::80 ${config.compress ? 'Compress:on' : 'Compress:off'}`,
-    '--defaultEntryPoints=http',
-  ];
-
-  // construct command
-  const Cmd = [
-    '-c',
-    '/dev/null',
-    '--docker',
-    '--docker.watch',
-    ...(config.letsencrypt ? letsencrypt : entrypoints),
-    ...(config.debug ? debug : []),
-    ...(config.traefikArgs || []),
-  ];
-
-  const Labels = {
-    'exoframe.deployment': 'exo-traefik',
-    'exoframe.user': 'admin',
-  };
-
-  const RestartPolicy = {
-    Name: 'on-failure',
-    MaximumRetryCount: 2,
-  };
-
-  // start traefik in docker
-  const container = await docker.createContainer({
-    Image: config.traefikImage,
-    name: config.traefikName,
-    Cmd,
-    Labels,
-    ExposedPorts: {
-      '80/tcp': {},
-      '443/tcp': {},
-    },
-    HostConfig: {
-      RestartPolicy,
-      Binds: ['/var/run/docker.sock:/var/run/docker.sock', `${acmePath}:/var/acme`],
-      PortBindings: {
-        '80/tcp': [{HostPort: '80'}],
-        '443/tcp': [{HostPort: '443'}],
-      },
-    },
-  });
-  // connect traefik to exoframe net
-  await exoNet.connect({
-    Container: container.id,
-  });
-  // start container
-  await container.start();
-  logger.info('Traefik instance started..');
+  // run traefik init
+  await initTraefik(exoNet);
 };