+++ src/routes/remove.js
@@ -2,6 +2,7 @@
 const docker = require('../docker/docker');
 const {removeContainer} = require('../docker/util');
 const {getPlugins} = require('../plugins');
+const {removeFunction} = require('exoframe-faas');
 const logger = require('../logger');
 
 // removal of normal containers
@@ -19,18 +20,32 @@ const removeUserContainer = async ({username, id, reply}) => {
     return;
   }
 
-  // if not found by name - try to find by project
-  const containers = allContainers.filter(
-    c => c.Labels['exoframe.user'] === username && c.Labels['exoframe.project'] === id
-  );
-  if (!containers.length) {
-    reply.code(404).send({error: 'Container not found!'});
+  // if not found by name - try to find by domain.
+  const containersByUrl = allContainers.filter(c => {
+    return (
+      c.Labels['exoframe.user'] === username &&
+      c.Labels[`traefik.http.routers.${c.Labels['exoframe.deployment']}.rule`] === `Host(\`${id}\`)`
+    );
+  });
+
+  if (containersByUrl.length) {
+    await Promise.all(containersByUrl.map(removeContainer));
+    reply.code(204).send('removed');
     return;
   }
-  // remove all
-  await Promise.all(containers.map(removeContainer));
-  // reply
-  reply.code(204).send('removed');
+
+  // if not found by name and url - try to find by project
+  const containersByProject = allContainers.filter(
+    c => c.Labels['exoframe.user'] === username && c.Labels['exoframe.project'] === id
+  );
+
+  if (containersByProject.length) {
+    await Promise.all(containersByProject.map(removeContainer));
+    reply.code(204).send('removed');
+    return;
+  }
+
+  reply.code(404).send({error: 'Container or function not found!'});
 };
 
 module.exports = fastify => {
@@ -41,6 +56,13 @@ module.exports = fastify => {
       // get username
       const {username} = request.user;
       const {id} = request.params;
+
+      // try and remove function
+      if (await removeFunction({id, username})) {
+        // reply
+        reply.code(204).send('removed');
+        return;
+      }
 
       // run remove via plugins if available
       const plugins = getPlugins();