+++ test/remove.test.js
@@ -8,7 +8,7 @@ const getPort = require('get-port');
 // our packages
 const authToken = require('./fixtures/authToken');
 const {startServer} = require('../src');
-const {pullImage} = require('../src/docker/init');
+const {pullImage} = require('../src/docker/util');
 const docker = require('../src/docker/docker');
 
 // options base
@@ -27,7 +27,7 @@ const projectName = 'rmtestproject';
 // fastify ref
 let fastify;
 
-const generateContainerConfig = ({name, username, project, baseName}) => ({
+const generateContainerConfig = ({name, username, project, baseName, url}) => ({
   Image: 'busybox:latest',
   Cmd: ['sh', '-c', 'sleep 1000'],
   name,
@@ -35,8 +35,7 @@ const generateContainerConfig = ({name, username, project, baseName}) => ({
     'exoframe.deployment': name,
     'exoframe.user': username,
     'exoframe.project': project,
-    'traefik.backend': baseName,
-    'traefik.frontend.rule': 'Host:test',
+    [`traefik.http.routers.${name}.rule`]: `Host(\`${url}\`)`,
   },
 });
 
@@ -57,6 +56,7 @@ beforeAll(async () => {
     username: 'admin',
     project: 'rmtest1',
     baseName: 'exo-admin-rmtest1',
+    url: 'test',
   });
   const container = await docker.createContainer(containerConfig);
   await container.start();
@@ -67,6 +67,7 @@ beforeAll(async () => {
     username: 'admin',
     project: projectName,
     baseName: 'exo-admin-rmtest2',
+    url: 'test',
   });
   const projectContainer1 = await docker.createContainer(prjContainerConfig1);
   await projectContainer1.start();
@@ -74,8 +75,9 @@ beforeAll(async () => {
   const prjContainerConfig2 = generateContainerConfig({
     name: 'rmtest3',
     username: 'admin',
-    project: projectName,
+    project: projectName + '_custom',
     baseName: 'exo-admin-rmtest3',
+    url: 'test.example.com',
   });
   const projectContainer2 = await docker.createContainer(prjContainerConfig2);
   await projectContainer2.start();
@@ -130,6 +132,17 @@ test('Should return error when removing nonexistent project', async done => {
   const result = JSON.parse(response.payload);
   // check response
   expect(response.statusCode).toEqual(404);
-  expect(result).toMatchObject({error: 'Container not found!'});
+  expect(result).toMatchObject({error: 'Container or function not found!'});
+  done();
+});
+
+test('Should remove by url', async done => {
+  const options = Object.assign({}, baseOptions, {
+    url: `/remove/test.example.com`,
+  });
+
+  const response = await fastify.inject(options);
+
+  expect(response.statusCode).toEqual(204);
   done();
 });